Split IO out of agent model module.
